Migrate to using ES6 modules.
This is generally invisible to the end-user; the Overture libraries are
still exported and still dropping everything in the `O` namespace. The
files should be trivially faster due to some reduced namespace lookups
(because rollup is cleverer than just plain `cat`) and the generated
minified + gzipped scripts are now around 3.4% smaller (and as we switch
to `let` and `const` it should improve further).

This also leaves the door open for more advanced optimisations later on,
with other libraries depending on individual pieces of Overture rather
than the library in its entirety, as dependencies are now tracked
precisely.

Things like the doc builder are still there unchanged for now. The build
scripts are also unchanged; it was easy to port Overture itself away
from those build scripts, but it will take much more effort to port
programs that used them to ES6 modules.

**Changes:** there are a few deliberate changes mixed in with this.

- `O.execute` (from the loader) has been removed as it’s working around
  bugs and inconsistencies from long-unsupported browsers in a way that
  doesn’t actually work in ES6 modules (so it’s easier to remove it than
  to fix it). If you had code that depended on it, (a) what on earth
  were you doing!? and (b) replace your `O.execute( code )` with the
  delightfully unconventional `( 1, eval )( code )`. Who said JavaScript
  didn’t have tuples? 😉

- `foundation/Binding.js` lost its support for running headlessly.
  It now requires `Element` to be in the globals or `O.Binding#connect`
  will blow up. This may sound like a big deal, but I believe that
  Overture depended on browser-specific functionality in a few other
  places as well, so if anyone actually wants headless support they’ll
  need to do a bit more to achieve it anyway.

- There is a general rule now of “one thing per file”. This affects
  internal code organisation and people using Overture via ES6 modules;
  people using the `O` global—everyone—are unaffected. A few things were
  shifted into new files because of this rule:

  - `sortByProperties`: `core/{Core → sortByProperties}.js`
  - `Event`: `foundation/{EventTarget → Event}.js`
  - `getFromPath`: `foundation/{ComputedProps → getFromPath}.js`
  - `parser`: `parser/{Parse → DateParser}.js` (it’d be *nice* to have
    it in `parser.js` but Windows paths are case-insensitive and
    `Parser.js` already exists; as it is, this namespace is anomalous in
    its naming and placement)
  - `StoreUndoManager`: `datastore/store/{UndoManager →
    StoreUndoManager}.js`
  - `formatKeyForPlatform`: `application/{GlobalKeyboardShortcut →
    formatKeyForPlatform}.js`
  - `MenuOptionView`: `views/controls/{MenuView → MenuOptionView}.js`
  - `MenuController`: `views/controls/{MenuView → MenuController}.js`
  - `ClearSearchButtonView`: `views/controls/{SearchTextView →
    ClearSearchButtonView}.js`
  - `CSSStyleAnimationController`: `animation/{CSSStyleAnimation →
    CSSStyleAnimationController}.js`

  Also at least one thing (uh oh) was shuffled between existing files:

  - `Record.attr`: `datastore/record/{RecordAttribute → Record}.js`

  Note that a few have *not* been shifted in cases where it might yet be
  good for them to be shifted; we can think about these later:

  - `Record.toMany` is still defined in `ToManyAttribute.js`
  - `Record.toOne` is still defined in `ToOneAttribute.js`
  - `Element.when` and `Element.unless` are defined in
    `views/collections/SwitchView.js` rather than `dom/Element.js`

**Side-effect imports:** we have quite a few cases where we extend
built-in types with new methods, like `Function#property`. Eventually I
hope to change most of these to standalone functions or decorators or
things like that, but for now we’re stuck with them. I’ve gone through
most of them and made sure that modules import the necessary other
modules for their side-effects. This is not an easy exercise for other
libraries, and is a grave stumbling-block to importing individual parts
of the Overture library, but I’ve mostly done it for Overture itself.
(These are the `import "../core/String.js";  // For String#capitalise`
imports.) I have a list, but this commit message is getting fairly long
and I think it’s relevant for the whole project of modernising Overture,
not just this commit, so I shan’t list them here. (I shall instead leave
them not checked in so that I can accidentally delete the list at the
drop of a hat. ’Cos that’s totally a great idea, y’know?)

**Honour roll:** one of the highlights of the exercise thus far is how
`datastore/record/Status.js` has embraced ES6 modules; this really helps
the rollup + uglify combination as it ends up able to inline all the
constants (which is a substantial fraction of the 3.4% savings). We’ll
see more of this sort of thing as we transition things to `let` and
`const`.

Viva la modernity!