New property O.UA.devicePixelRatio
Use this observable property instead of `window.devicePixelRatio`.

`devicePixelRatio` can change if the browser zoom level changes, or the
browser window is moved between low- and high-DPI screens and suchlike.
But bear in mind with this approach to it that you get changes like from
1 to 1.0909091234207153 for zooming in one notch; this isn’t just for 1×
to 2× and the likes.

And if you’re using `devicePixelRatio` for images, well, you should
probably use `srcset` or something similar instead (that way the browser
takes care of it and the old image can be displayed while it loads the
new image).

So in reality this property probably isn’t actually as useful as you
thought, because you were probably using devicePixelRatio incorrectly.