Make __meta__ a Symbol
This makes it more private and keeps it out of regular iteration.

This is really inconvenient for debugging records in particular;
in browser dev tools you could formerly expand `__meta__` (pretty much
always the first item) and then expand `cache` to look at the values for
all the properties on the record. I aim to make this more convenient
again in the future with some of the more radical changes. And with an
Overture dev tools panel or some such thing. (Debugging Overture objects
is generally pretty terrible right now anyway.)