Initialise special properties lazily
Formerly, special properties were initialised immediately in the
`Class()` or `mixin()` call.

Now, they are initialised when `meta()` is first called on the object
(whether it be an instance or a class prototype), which is normally
going to be when you create the first instance of a class.

Classes with no special properties will have an own __meta__ property,
but it will be `undefined` rather than a Metadata instance. This is a
microoptimisation. Class *instances* always have metadata (the
`O.Object` constructor ensures this).

## Using `class` syntax

Thanks to this, `O.Class` is no longer necessary. The following:

```javascript
const Foo = Class({
    Extends: Bar,
    Mixins: [ MixA, MixB ],

    init: function () {
        Foo.parent.init.apply( this, arguments );
    },

    foo: 42,

    bar: function () {
    }.property( 'foo' ),

    baz: function () {
    },
});
```

can now be written this way if you desire:

```javascript
class Foo extends Bar {
    constructor ( ...args ) {
        super( ...args );
    }
}

Object.assign( Foo.prototype, MixA, MixB, {
    foo: 42,

    bar: function () {
    }.property( 'foo' ),

    baz () {
    }
});
```

Because of how ugly adding non-methods to the prototype is with the
`class` syntax, `O.Class` is not deprecated at this stage. However,
using `class` syntax gets you some nice advantages like ideal handling
of super methods and an actual name on the class (which is very nice for
debugging).

(Future specs that may make class syntax nicer: field syntax, which puts
values on instances—though not the prototype; and decorators for all the
Function.prototype methods we inject.)

## Breaking change

This is a theoretical breaking change for several reasons:

- Classes that do not extend `O.Object` but contained special properties
  will now be broken unless `O.meta()` is called on the class prototype
  or an instance of the class. I do not consider this an important
  breakage as classes with special properties should all extend
  `O.Object`.

- Adding *or overriding* special properties with `mixin()` after
  `meta()` is called on the class prototype or an instance is created
  used to mostly work (though in the right combination of circumstances
  it would blow up), but now it won’t work at all. Again, this is one of
  those things that you shouldn’t have been doing and was risky
  behaviour, so I’m not too worried about it.

In short, this is technically a breaking change, but you should be
heartily ashamed of any code that gets broken by this change, because
it was wrong anyway and merely hadn’t blown up in your face *yet*.

(OK, I admit it: I won’t be surprised if there are actually one or two
breakages, though I don’t expect any; the suspect behaviour was truly
quite subtle in some situations.)

## Deprecations

`O.mixin()` is now deprecated since it is no longer performing the magic
of special property initialisation.

`O.Class` is not deprecated, but you might want to look at ES6 class
syntax all the same.