WIP: Bring ObservableArray into an ES6 world
It now uses ES6 class syntax, and `length` and `[]` were changed from
magic properties to regular ES5 descriptors with getters and setters.
When doing it this way, basically the only thing that needs to be done
is a `propertyDidChange` call.

FIXME: `[]` should be cached, because
… HMM. SHOULD IT BE?
As part of this change `[]` is no longer a cached property. It made no
sense for it to be; `this[ __meta__ ].cache[ '[]' ]` is hardly going to
be faster than `this._array`!

I also added faster implementations of the `Enumerable` methods.

There is an important potential breaking changes in here: the
ObservableArray constructor used to call the super constructor *after*
setting `_array`; in ES6 classes, super() must be called before `this`
is touched, so the super() call happens before `_array` is set.
Therefore, when bindings are initialised, *the ObservableArray is
unusable*. Anything that tries to get something out of the array at that
point *will blow up*. I don’t know if this is an issue or not yet. If it
is, I’ll think of something (I don’t know what yet).