WIP: Make ObservableArray use getters and setters
WIP because of the notes of breaking changes and differences in
behaviour.

Concerning the breaking change of constructor ordering, Neil made the
suggestion that `super( { _array: array, … }, ...mixins );` would work
to set overrideable properties. That is quite an elegant solution to the
first part of the problem, though it doesn’t play nicely with property
initialiser syntax (that is, `class { foo = 42; }`) as used in
TypeScript and a stage 3 proposal for ECMAScript. What we shall do
remains to be seen.