Deprecate View#positioning
It’s still there, but should be avoided in favour of adding `position`
to the View#layout property object. It will disappear from Overture when
we no longer need it at FastMail, if not earlier.

This may break some things, if they overrode `positioning` specifically;
although View itself still supports positioning, certain of its
subclasses (those changed in this commit, no more) are incidentally
overriding that value in their definition of `layout`.