WIP: Update Todo example for Overture changes
The RPC language which was a precursor of what is now being standardised
in JMAP has now been replaced with JMAP. This means changing getFoo to
Foo/get and things like that. Multi-account support has necessitated
certain changes in that as well.

Connection.js has been updated to the current state of that file from
the jmap-js project.

fixtures.js has some additional changes in line with what we’ve done in
the equivalent stuff in FastMail and/or Topicbox.

This also includes some style changes required due to DOM structure
changes in ToolbarView and I think ClearSearchButtonView too.

KNOWN REGRESSIONS from the last-working version of Todo on
overturejs.com: toggling isComplete commits, but does not create a new
undo checkpoint. Not sure why yet. I’m guessing an undo manager bug.